backend/requirements.txt
fastapi
uvicorn
pymongo
spacy
pdfplumber
python-docx
requests
beautifulsoup4
apscheduler
openpyxl
reportlab
python-multipart
python-jose[cryptography]
python-dotenv
